Move files in prep op unit-testing and gradle
