+++ src/compilerModule/src/main/java/module-info.java
@@ -1,0 +1,8 @@
+module compilerModule {
+    requires jdk.compiler;
+    requires java.base;
+    requires java.desktop;
+    requires java.management;
+    requires java.sql;
+    requires java.xml;
+}
