--- src/compilerModule/src/module-info.java
@@ -1,8 +1,0 @@
-module compilerModule {
-    requires jdk.compiler;
-    requires java.base;
-    requires java.desktop;
-    requires java.management;
-    requires java.sql;
-    requires java.xml;
-}
